+++ config/webhook-client.php
@@ -44,6 +44,16 @@ return [
             'webhook_model' => \Spatie\WebhookClient\Models\WebhookCall::class,
 
             /*
+             * In this array, you can pass the headers that should be store on
+             * the webhook call model when a webhook comes in.
+             *
+             * To store all headers, set this value to `*`.
+             */
+            'store_headers' => [
+
+            ],
+
+            /*
              * The class name of the job that will process the webhook request.
              *
              * This should be set to a class that extends \Spatie\WebhookClient\ProcessWebhookJob.