+++ src/Models/WebhookCall.php
@@ -20,11 +20,31 @@ class WebhookCall extends Model
 
     public static function storeWebhook(WebhookConfig $config, Request $request): WebhookCall
     {
+        $headers = self::headersToStore($config, $request);
+
         return self::create([
             'name' => $config->name,
-            'headers' => $request->headers->all(),
+            'headers' => $headers,
             'payload' => $request->input(),
         ]);
+    }
+
+    public static function headersToStore(WebhookConfig $config, Request $request): array
+    {
+        $headerNamesToStore = $config->storeHeaders;
+
+        if ($headerNamesToStore === '*') {
+            return $request->headers->all();
+        }
+
+        $headerNamesToStore = array_map(
+            fn(string $headerName) => strtolower($headerName),
+            $headerNamesToStore,
+        );
+
+        return collect($request->headers->all())
+            ->filter(fn(array $headerValue, string $headerName) => in_array($headerName, $headerNamesToStore))
+            ->toArray();
     }
 
     public function headerBag(): HeaderBag